requirements.txt
pdf2image
Pillow
google-generativeai
